requirements.txt
lasagna-ai==0.11.1
twilio-phone-calls==1.3
beautifulsoup4==4.13.3
tqdm
pydantic
